QmailExit: print the log "no matter what"
OK, should trigger on "die" or missing qexit at end of program.

But will not trigger on exec, nor even on system
$original_qmail_remote, as that probably prints \0 and then the note
printing is swallowed.

Should we not use exec to run orig qmail-remote, but wrap it while
capturing its output and then extending that?

What a messy design is this?